Add preshared key generation for authentication
This allows for secure authentication for IPC between the parent
and child process. The parent is given a hash of the key, and the
child is passed the key. The child should authenticate, and drop
this key as fast as possible.

The key is a [u8; 64], and is piped to the child via stdin.
This should be the most secure way to pass the key to the child as
the pipe will store the key in kernel memory, rather than user
space.